+++ core/src/main/scala/org/bitcoins/core/psbt/PSBT.scala
@@ -1,21 +1,19 @@
 package org.bitcoins.core.psbt
 
-import org.bitcoins.core.byteVectorOrdering
 import org.bitcoins.core.crypto._
 import org.bitcoins.core.hd.BIP32Path
 import org.bitcoins.core.number.UInt32
 import org.bitcoins.core.policy.Policy
-import org.bitcoins.core.protocol.script.{P2WSHWitnessV0, _}
+import org.bitcoins.core.protocol.NetworkElement
+import org.bitcoins.core.protocol.script._
 import org.bitcoins.core.protocol.transaction._
-import org.bitcoins.core.protocol.{CompactSizeUInt, NetworkElement}
 import org.bitcoins.core.script.PreExecutionScriptProgram
 import org.bitcoins.core.script.crypto.HashType
 import org.bitcoins.core.script.interpreter.ScriptInterpreter
 import org.bitcoins.core.script.result.ScriptOk
-import org.bitcoins.core.serializers.script.RawScriptWitnessParser
-import org.bitcoins.core.util.{CryptoUtil, Factory}
+import org.bitcoins.core.util.Factory
 import org.bitcoins.core.wallet.builder.BitcoinTxBuilder
-import org.bitcoins.core.wallet.signer.{BitcoinSigner, BitcoinSignerSingle}
+import org.bitcoins.core.wallet.signer.BitcoinSignerSingle
 import org.bitcoins.core.wallet.utxo._
 import scodec.bits._
 
@@ -692,1229 +690,4 @@ object PSBT extends Factory[PSBT] {
       PSBT(globalMap, inputMaps, outputMaps)
     }
   }
-}
-
-sealed trait PSBTRecord extends NetworkElement {
-
-  type KeyId <: PSBTKeyId
-
-  def key: ByteVector
-
-  def value: ByteVector
-
-  def bytes: ByteVector = {
-    val keySize = CompactSizeUInt.calc(key)
-    val valueSize = CompactSizeUInt.calc(value)
-
-    keySize.bytes ++ key ++ valueSize.bytes ++ value
-  }
-}
-
-object PSBTRecord {
-
-  def fromBytes(bytes: ByteVector): (ByteVector, ByteVector) = {
-    val keyCmpctUInt = CompactSizeUInt.parse(bytes)
-
-    if (keyCmpctUInt.toLong == 0) {
-      (ByteVector.empty, ByteVector.empty)
-    } else {
-      val key = bytes.drop(keyCmpctUInt.size).take(keyCmpctUInt.toLong)
-      val valueBytes = bytes.drop(keyCmpctUInt.size + keyCmpctUInt.toLong)
-      val valueCmpctUInt = CompactSizeUInt.parse(valueBytes)
-      val value = valueBytes
-        .drop(valueCmpctUInt.size)
-        .take(valueCmpctUInt.toLong)
-
-      (key, value)
-    }
-  }
-}
-
-sealed trait GlobalPSBTRecord extends PSBTRecord {
-  override type KeyId <: PSBTGlobalKeyId
-}
-
-object GlobalPSBTRecord extends Factory[GlobalPSBTRecord] {
-  import org.bitcoins.core.psbt.PSBTGlobalKeyId._
-  case class UnsignedTransaction(transaction: BaseTransaction)
-      extends GlobalPSBTRecord {
-    require(
-      transaction.inputs.forall(_.scriptSignature == EmptyScriptSignature),
-      s"All ScriptSignatures must be empty, got $transaction")
-
-    override type KeyId = UnsignedTransactionKeyId.type
-    override val key: ByteVector = ByteVector(UnsignedTransactionKeyId.byte)
-    override val value: ByteVector = transaction.bytes
-  }
-
-  case class XPubKey(
-      xpub: ExtPublicKey,
-      masterFingerprint: ByteVector,
-      derivationPath: BIP32Path)
-      extends GlobalPSBTRecord {
-    require(
-      derivationPath.path.length == xpub.depth.toInt,
-      s"Derivation path length does not match xpubkey depth, difference: ${derivationPath.path.length - xpub.depth.toInt}"
-    )
-    require(
-      masterFingerprint.length == 4,
-      s"Master key fingerprints are 4 bytes long, got: $masterFingerprint")
-
-    override type KeyId = XPubKeyKeyId.type
-    override val key: ByteVector = ByteVector(XPubKeyKeyId.byte) ++ xpub.bytes
-    override val value: ByteVector = derivationPath.path
-      .foldLeft(masterFingerprint)(_ ++ _.toUInt32.bytesLE)
-  }
-
-  case class Version(version: UInt32) extends GlobalPSBTRecord {
-    override type KeyId = VersionKeyId.type
-    override val key: ByteVector = ByteVector(VersionKeyId.byte)
-    override val value: ByteVector = version.bytesLE
-  }
-
-  case class Unknown(key: ByteVector, value: ByteVector)
-      extends GlobalPSBTRecord {
-    override type KeyId = UnknownKeyId.type
-    private val keyId = PSBTGlobalKeyId.fromBytes(key)
-    require(keyId == UnknownKeyId,
-            s"Cannot make an Unknown record with a $keyId")
-  }
-
-  override def fromBytes(bytes: ByteVector): GlobalPSBTRecord = {
-    import org.bitcoins.core.psbt.PSBTGlobalKeyId._
-
-    val (key, value) = PSBTRecord.fromBytes(bytes)
-    PSBTGlobalKeyId.fromByte(key.head) match {
-      case UnsignedTransactionKeyId =>
-        require(key.size == 1,
-                s"The key must only contain the 1 byte type, got: ${key.size}")
-
-        UnsignedTransaction(BaseTransaction.fromBytes(value))
-      case XPubKeyKeyId =>
-        val xpub = ExtPublicKey.fromBytes(key.tail.take(78))
-        val fingerprint = value.take(4)
-        val path = BIP32Path.fromBytesLE(value.drop(4))
-        XPubKey(xpub, fingerprint, path)
-      case VersionKeyId =>
-        require(key.size == 1,
-                s"The key must only contain the 1 byte type, got: ${key.size}")
-
-        val version = UInt32.fromBytesLE(value)
-
-        require(PSBT.knownVersions.contains(version),
-                s"Unknown version number given: $version")
-        Version(version)
-      case UnknownKeyId =>
-        GlobalPSBTRecord.Unknown(key, value)
-    }
-  }
-}
-
-sealed trait InputPSBTRecord extends PSBTRecord {
-  override type KeyId <: PSBTInputKeyId
-}
-
-object InputPSBTRecord extends Factory[InputPSBTRecord] {
-  import org.bitcoins.core.psbt.PSBTInputKeyId._
-  case class NonWitnessOrUnknownUTXO(transactionSpent: Transaction)
-      extends InputPSBTRecord {
-    override type KeyId = NonWitnessUTXOKeyId.type
-    override val key: ByteVector = ByteVector(NonWitnessUTXOKeyId.byte)
-    override val value: ByteVector = transactionSpent.bytes
-  }
-
-  case class WitnessUTXO(witnessUTXO: TransactionOutput)
-      extends InputPSBTRecord {
-    override type KeyId = WitnessUTXOKeyId.type
-    override val key: ByteVector = ByteVector(WitnessUTXOKeyId.byte)
-    override val value: ByteVector = witnessUTXO.bytes
-  }
-
-  case class PartialSignature(
-      pubKey: ECPublicKey,
-      signature: ECDigitalSignature)
-      extends InputPSBTRecord {
-    require(pubKey.size == 33, s"pubKey must be 33 bytes, got: ${pubKey.size}")
-
-    override type KeyId = PartialSignatureKeyId.type
-    override val key: ByteVector = ByteVector(PartialSignatureKeyId.byte) ++ pubKey.bytes
-    override val value: ByteVector = signature.bytes
-  }
-
-  case class SigHashType(hashType: HashType) extends InputPSBTRecord {
-    override type KeyId = SigHashTypeKeyId.type
-    override val key: ByteVector = ByteVector(SigHashTypeKeyId.byte)
-    override val value: ByteVector = hashType.num.bytesLE
-  }
-
-  case class RedeemScript(redeemScript: ScriptPubKey) extends InputPSBTRecord {
-    override type KeyId = RedeemScriptKeyId.type
-    override val key: ByteVector = ByteVector(RedeemScriptKeyId.byte)
-    override val value: ByteVector = redeemScript.asmBytes
-  }
-
-  case class WitnessScript(witnessScript: RawScriptPubKey)
-      extends InputPSBTRecord {
-    override type KeyId = WitnessScriptKeyId.type
-    override val key: ByteVector = ByteVector(WitnessScriptKeyId.byte)
-    override val value: ByteVector = witnessScript.asmBytes
-  }
-
-  case class BIP32DerivationPath(
-      pubKey: ECPublicKey,
-      masterFingerprint: ByteVector,
-      path: BIP32Path)
-      extends InputPSBTRecord {
-    require(pubKey.size == 33, s"pubKey must be 33 bytes, got: ${pubKey.size}")
-
-    override type KeyId = BIP32DerivationPathKeyId.type
-    override val key: ByteVector = ByteVector(BIP32DerivationPathKeyId.byte) ++ pubKey.bytes
-    override val value: ByteVector =
-      path.path.foldLeft(masterFingerprint)(_ ++ _.toUInt32.bytesLE)
-  }
-
-  case class FinalizedScriptSig(scriptSig: ScriptSignature)
-      extends InputPSBTRecord {
-    override type KeyId = FinalizedScriptSigKeyId.type
-    override val key: ByteVector = ByteVector(FinalizedScriptSigKeyId.byte)
-    override val value: ByteVector = scriptSig.asmBytes
-  }
-
-  case class FinalizedScriptWitness(scriptWitness: ScriptWitness)
-      extends InputPSBTRecord {
-    override type KeyId = FinalizedScriptWitnessKeyId.type
-    override val key: ByteVector = ByteVector(FinalizedScriptWitnessKeyId.byte)
-    override val value: ByteVector = scriptWitness.bytes
-  }
-
-  case class ProofOfReservesCommitment(porCommitment: ByteVector)
-      extends InputPSBTRecord {
-    override type KeyId = ProofOfReservesCommitmentKeyId.type
-    override val key: ByteVector = ByteVector(
-      ProofOfReservesCommitmentKeyId.byte)
-    override val value: ByteVector = porCommitment
-  }
-
-  case class Unknown(key: ByteVector, value: ByteVector)
-      extends InputPSBTRecord {
-    override type KeyId = UnknownKeyId.type
-    private val keyId = PSBTInputKeyId.fromBytes(key)
-    require(keyId == UnknownKeyId,
-            s"Cannot make an Unknown record with a $keyId")
-  }
-
-  override def fromBytes(bytes: ByteVector): InputPSBTRecord = {
-    import org.bitcoins.core.psbt.PSBTInputKeyId._
-
-    val (key, value) = PSBTRecord.fromBytes(bytes)
-    PSBTInputKeyId.fromByte(key.head) match {
-      case NonWitnessUTXOKeyId =>
-        require(key.size == 1,
-                s"The key must only contain the 1 byte type, got: ${key.size}")
-
-        NonWitnessOrUnknownUTXO(Transaction(value))
-      case WitnessUTXOKeyId =>
-        require(key.size == 1,
-                s"The key must only contain the 1 byte type, got: ${key.size}")
-
-        WitnessUTXO(TransactionOutput.fromBytes(value))
-      case PartialSignatureKeyId =>
-        val pubKey = ECPublicKey(key.tail)
-        val sig = ECDigitalSignature(value)
-        PartialSignature(pubKey, sig)
-      case SigHashTypeKeyId =>
-        require(key.size == 1,
-                s"The key must only contain the 1 byte type, got: ${key.size}")
-
-        SigHashType(HashType.fromBytesLE(value))
-      case PSBTInputKeyId.RedeemScriptKeyId =>
-        require(key.size == 1,
-                s"The key must only contain the 1 byte type, got: ${key.size}")
-
-        InputPSBTRecord.RedeemScript(ScriptPubKey.fromAsmBytes(value))
-      case PSBTInputKeyId.WitnessScriptKeyId =>
-        require(key.size == 1,
-                s"The key must only contain the 1 byte type, got: ${key.size}")
-
-        InputPSBTRecord.WitnessScript(RawScriptPubKey.fromAsmBytes(value))
-      case PSBTInputKeyId.BIP32DerivationPathKeyId =>
-        val pubKey = ECPublicKey(key.tail)
-        val fingerprint = value.take(4)
-        val path = BIP32Path.fromBytesLE(value.drop(4))
-        InputPSBTRecord.BIP32DerivationPath(pubKey, fingerprint, path)
-      case FinalizedScriptSigKeyId =>
-        require(key.size == 1,
-                s"The key must only contain the 1 byte type, got: ${key.size}")
-
-        val sig = ScriptSignature.fromAsmBytes(value)
-        FinalizedScriptSig(sig)
-      case FinalizedScriptWitnessKeyId =>
-        require(key.size == 1,
-                s"The key must only contain the 1 byte type, got: ${key.size}")
-
-        FinalizedScriptWitness(RawScriptWitnessParser.read(value))
-      case ProofOfReservesCommitmentKeyId =>
-        ProofOfReservesCommitment(value)
-      case UnknownKeyId =>
-        InputPSBTRecord.Unknown(key, value)
-    }
-  }
-}
-
-sealed trait OutputPSBTRecord extends PSBTRecord {
-  override type KeyId <: PSBTOutputKeyId
-}
-
-object OutputPSBTRecord extends Factory[OutputPSBTRecord] {
-  import org.bitcoins.core.psbt.PSBTOutputKeyId._
-
-  case class RedeemScript(redeemScript: ScriptPubKey) extends OutputPSBTRecord {
-    override type KeyId = RedeemScriptKeyId.type
-    override val key: ByteVector = ByteVector(RedeemScriptKeyId.byte)
-    override val value: ByteVector = redeemScript.asmBytes
-  }
-
-  case class WitnessScript(witnessScript: ScriptPubKey)
-      extends OutputPSBTRecord {
-    override type KeyId = WitnessScriptKeyId.type
-    override val key: ByteVector = ByteVector(WitnessScriptKeyId.byte)
-    override val value: ByteVector = witnessScript.asmBytes
-  }
-
-  case class BIP32DerivationPath(
-      pubKey: ECPublicKey,
-      masterFingerprint: ByteVector,
-      path: BIP32Path)
-      extends OutputPSBTRecord {
-    require(pubKey.size == 33, s"pubKey must be 33 bytes, got: ${pubKey.size}")
-
-    override type KeyId = BIP32DerivationPathKeyId.type
-    override val key: ByteVector = ByteVector(BIP32DerivationPathKeyId.byte) ++ pubKey.bytes
-    override val value: ByteVector =
-      path.path.foldLeft(masterFingerprint)(_ ++ _.toUInt32.bytesLE)
-  }
-
-  case class Unknown(key: ByteVector, value: ByteVector)
-      extends OutputPSBTRecord {
-    override type KeyId = UnknownKeyId.type
-    private val keyId = PSBTOutputKeyId.fromBytes(key)
-    require(keyId == UnknownKeyId,
-            s"Cannot make an Unknown record with a $keyId")
-  }
-
-  override def fromBytes(bytes: ByteVector): OutputPSBTRecord = {
-    val (key, value) = PSBTRecord.fromBytes(bytes)
-    PSBTOutputKeyId.fromByte(key.head) match {
-      case PSBTOutputKeyId.RedeemScriptKeyId =>
-        require(key.size == 1,
-                s"The key must only contain the 1 byte type, got: ${key.size}")
-
-        OutputPSBTRecord.RedeemScript(ScriptPubKey.fromAsmBytes(value))
-      case PSBTOutputKeyId.WitnessScriptKeyId =>
-        require(key.size == 1,
-                s"The key must only contain the 1 byte type, got: ${key.size}")
-
-        OutputPSBTRecord.WitnessScript(ScriptPubKey.fromAsmBytes(value))
-      case PSBTOutputKeyId.BIP32DerivationPathKeyId =>
-        val pubKey = ECPublicKey(key.tail)
-        val fingerprint = value.take(4)
-        val path = BIP32Path.fromBytesLE(value.drop(4))
-
-        OutputPSBTRecord.BIP32DerivationPath(pubKey, fingerprint, path)
-      case UnknownKeyId =>
-        OutputPSBTRecord.Unknown(key, value)
-    }
-  }
-}
-
-/**
-  * A PSBTKeyId refers to the first byte of a key that signifies which kind of key-value map
-  * is in a given PSBTRecord
-  */
-sealed trait PSBTKeyId {
-  def byte: Byte
-  type RecordType <: PSBTRecord
-}
-
-sealed trait PSBTKeyIdFactory[KeyIdType <: PSBTKeyId]
-    extends Factory[KeyIdType] {
-  def fromByte(byte: Byte): KeyIdType
-  def unknownKey: KeyIdType
-
-  override def fromBytes(key: ByteVector): KeyIdType = {
-    if (key.isEmpty) {
-      unknownKey
-    } else {
-      fromByte(key.head)
-    }
-  }
-}
-
-sealed trait PSBTGlobalKeyId extends PSBTKeyId {
-  type RecordType <: GlobalPSBTRecord
-}
-
-object PSBTGlobalKeyId extends PSBTKeyIdFactory[PSBTGlobalKeyId] {
-
-  override def fromByte(byte: Byte): PSBTGlobalKeyId = byte match {
-    case UnsignedTransactionKeyId.byte => UnsignedTransactionKeyId
-    case XPubKeyKeyId.byte             => XPubKeyKeyId
-    case VersionKeyId.byte             => VersionKeyId
-    case _: Byte                       => UnknownKeyId
-  }
-
-  override def unknownKey: PSBTGlobalKeyId = UnknownKeyId
-
-  final case object UnsignedTransactionKeyId extends PSBTGlobalKeyId {
-    override val byte: Byte = 0x00.byteValue
-    type RecordType = GlobalPSBTRecord.UnsignedTransaction
-  }
-
-  final case object XPubKeyKeyId extends PSBTGlobalKeyId {
-    override val byte: Byte = 0x01.byteValue
-    type RecordType = GlobalPSBTRecord.XPubKey
-  }
-
-  final case object VersionKeyId extends PSBTGlobalKeyId {
-    override val byte: Byte = 0xFB.byteValue
-    type RecordType = GlobalPSBTRecord.Version
-  }
-
-  final case object UnknownKeyId extends PSBTGlobalKeyId {
-    override val byte: Byte = Byte.MaxValue
-    type RecordType = GlobalPSBTRecord.Unknown
-  }
-}
-
-sealed trait PSBTInputKeyId extends PSBTKeyId {
-  type RecordType <: InputPSBTRecord
-}
-
-object PSBTInputKeyId extends PSBTKeyIdFactory[PSBTInputKeyId] {
-
-  override def unknownKey: PSBTInputKeyId = UnknownKeyId
-
-  override def fromByte(byte: Byte): PSBTInputKeyId = byte match {
-    case NonWitnessUTXOKeyId.byte            => NonWitnessUTXOKeyId
-    case WitnessUTXOKeyId.byte               => WitnessUTXOKeyId
-    case PartialSignatureKeyId.byte          => PartialSignatureKeyId
-    case SigHashTypeKeyId.byte               => SigHashTypeKeyId
-    case RedeemScriptKeyId.byte              => RedeemScriptKeyId
-    case WitnessScriptKeyId.byte             => WitnessScriptKeyId
-    case BIP32DerivationPathKeyId.byte       => BIP32DerivationPathKeyId
-    case FinalizedScriptSigKeyId.byte        => FinalizedScriptSigKeyId
-    case FinalizedScriptWitnessKeyId.byte    => FinalizedScriptWitnessKeyId
-    case ProofOfReservesCommitmentKeyId.byte => ProofOfReservesCommitmentKeyId
-    case _: Byte                             => UnknownKeyId
-
-  }
-
-  final case object NonWitnessUTXOKeyId extends PSBTInputKeyId {
-    override val byte: Byte = 0x00.byteValue
-    type RecordType = InputPSBTRecord.NonWitnessOrUnknownUTXO
-  }
-
-  final case object WitnessUTXOKeyId extends PSBTInputKeyId {
-    override val byte: Byte = 0x01.byteValue
-    type RecordType = InputPSBTRecord.WitnessUTXO
-  }
-
-  final case object PartialSignatureKeyId extends PSBTInputKeyId {
-    override val byte: Byte = 0x02.byteValue
-    type RecordType = InputPSBTRecord.PartialSignature
-  }
-
-  final case object SigHashTypeKeyId extends PSBTInputKeyId {
-    override val byte: Byte = 0x03.byteValue
-    type RecordType = InputPSBTRecord.SigHashType
-  }
-
-  final case object RedeemScriptKeyId extends PSBTInputKeyId {
-    override val byte: Byte = 0x04.byteValue
-    type RecordType = InputPSBTRecord.RedeemScript
-  }
-
-  final case object WitnessScriptKeyId extends PSBTInputKeyId {
-    override val byte: Byte = 0x05.byteValue
-    type RecordType = InputPSBTRecord.WitnessScript
-  }
-
-  final case object BIP32DerivationPathKeyId extends PSBTInputKeyId {
-    override val byte: Byte = 0x06.byteValue
-    type RecordType = InputPSBTRecord.BIP32DerivationPath
-  }
-
-  final case object FinalizedScriptSigKeyId extends PSBTInputKeyId {
-    override val byte: Byte = 0x07.byteValue
-    type RecordType = InputPSBTRecord.FinalizedScriptSig
-  }
-
-  final case object FinalizedScriptWitnessKeyId extends PSBTInputKeyId {
-    override val byte: Byte = 0x08.byteValue
-    type RecordType = InputPSBTRecord.FinalizedScriptWitness
-  }
-
-  final case object ProofOfReservesCommitmentKeyId extends PSBTInputKeyId {
-    override val byte: Byte = 0x09.byteValue
-    type RecordType = InputPSBTRecord.ProofOfReservesCommitment
-  }
-
-  final case object UnknownKeyId extends PSBTInputKeyId {
-    override val byte: Byte = Byte.MaxValue
-    type RecordType = InputPSBTRecord.Unknown
-  }
-}
-
-sealed trait PSBTOutputKeyId extends PSBTKeyId {
-  type RecordType <: OutputPSBTRecord
-}
-
-object PSBTOutputKeyId extends PSBTKeyIdFactory[PSBTOutputKeyId] {
-
-  override def unknownKey: PSBTOutputKeyId = UnknownKeyId
-
-  override def fromByte(byte: Byte): PSBTOutputKeyId = byte match {
-    case RedeemScriptKeyId.byte        => RedeemScriptKeyId
-    case WitnessScriptKeyId.byte       => WitnessScriptKeyId
-    case BIP32DerivationPathKeyId.byte => BIP32DerivationPathKeyId
-    case _: Byte                       => UnknownKeyId
-  }
-
-  final case object RedeemScriptKeyId extends PSBTOutputKeyId {
-    override val byte: Byte = 0x00.byteValue
-    type RecordType = OutputPSBTRecord.RedeemScript
-  }
-
-  final case object WitnessScriptKeyId extends PSBTOutputKeyId {
-    override val byte: Byte = 0x01.byteValue
-    type RecordType = OutputPSBTRecord.WitnessScript
-  }
-
-  final case object BIP32DerivationPathKeyId extends PSBTOutputKeyId {
-    override val byte: Byte = 0x02.byteValue
-    type RecordType = OutputPSBTRecord.BIP32DerivationPath
-  }
-
-  final case object UnknownKeyId extends PSBTOutputKeyId {
-    override val byte: Byte = Byte.MaxValue
-    type RecordType = OutputPSBTRecord.Unknown
-  }
-}
-
-sealed trait PSBTMap[+RecordType <: PSBTRecord] extends NetworkElement {
-  require(elements.map(_.key).groupBy(identity).values.forall(_.length == 1),
-          "All keys must be unique.")
-
-  def elements: Vector[RecordType]
-
-  def bytes: ByteVector =
-    elements
-      .sortBy(_.key)
-      .foldLeft(ByteVector.empty)(_ ++ _.bytes) :+ PSBTMap.separatorByte
-
-  protected def getRecords[KeyIdType <: PSBTKeyId](
-      key: KeyIdType,
-      keyIdFactory: PSBTKeyIdFactory[KeyIdType]): Vector[key.RecordType] = {
-    elements
-      .filter(element => keyIdFactory.fromByte(element.key.head) == key)
-      .asInstanceOf[Vector[key.RecordType]]
-  }
-
-  protected def filterRecords[KeyIdType <: PSBTKeyId](
-      key: KeyIdType,
-      keyIdFactory: PSBTKeyIdFactory[KeyIdType]): Vector[RecordType] = {
-    elements
-      .filterNot(element => keyIdFactory.fromByte(element.key.head) == key)
-  }
-}
-
-object PSBTMap {
-  final val separatorByte: Byte = 0x00.byteValue
-}
-
-sealed trait PSBTMapFactory[
-    RecordType <: PSBTRecord, MapType <: PSBTMap[RecordType]]
-    extends Factory[MapType] {
-  def recordFactory: Factory[RecordType]
-
-  def constructMap(elements: Vector[RecordType]): MapType
-
-  lazy val empty: MapType = constructMap(Vector.empty)
-
-  override def fromBytes(bytes: ByteVector): MapType = {
-    @tailrec
-    def loop(
-        remainingBytes: ByteVector,
-        accum: Vector[RecordType]): Vector[RecordType] = {
-      if (remainingBytes.head == PSBTMap.separatorByte) {
-        accum
-      } else {
-        val record = recordFactory.fromBytes(remainingBytes)
-        val next = remainingBytes.drop(record.bytes.size)
-
-        loop(next, accum :+ record)
-      }
-    }
-
-    constructMap(loop(bytes, Vector.empty))
-  }
-}
-
-case class GlobalPSBTMap(elements: Vector[GlobalPSBTRecord])
-    extends PSBTMap[GlobalPSBTRecord] {
-  import org.bitcoins.core.psbt.GlobalPSBTRecord._
-  import org.bitcoins.core.psbt.PSBTGlobalKeyId._
-  require(getRecords(UnsignedTransactionKeyId).nonEmpty,
-          "A GlobalPSBTMap must have a Unsigned Transaction")
-
-  def unsignedTransaction: UnsignedTransaction = {
-    getRecords(UnsignedTransactionKeyId).head
-  }
-
-  def extendedPublicKeys: Vector[XPubKey] = {
-    getRecords(XPubKeyKeyId)
-  }
-
-  def version: Version = {
-    getRecords(VersionKeyId).headOption.getOrElse(Version(UInt32.zero))
-  }
-
-  private def getRecords(key: PSBTGlobalKeyId): Vector[key.RecordType] = {
-    super.getRecords(key, PSBTGlobalKeyId)
-  }
-
-  def filterRecords(key: PSBTGlobalKeyId): Vector[GlobalPSBTRecord] = {
-    super.filterRecords(key, PSBTGlobalKeyId)
-  }
-
-  /**
-    * Takes another GlobalPSBTMap and adds all records that are not contained in this GlobalPSBTMap
-    * @param other GlobalPSBTMap to be combined with
-    * @return A GlobalPSBTMap with the combined data of the two GlobalPSBTMaps
-    */
-  def combine(other: GlobalPSBTMap): GlobalPSBTMap = {
-    require(
-      this
-        .getRecords(UnsignedTransactionKeyId)
-        .head == other.getRecords(UnsignedTransactionKeyId).head,
-      "Cannot combine GlobalPSBTMaps with different unsigned transactions"
-    )
-    // We must keep the highest version number
-    // https://github.com/bitcoin/bips/blob/master/bip-0174.mediawiki#version-numbers
-    if (this.version.version > other.version.version) {
-      GlobalPSBTMap(
-        (this.elements ++ other.filterRecords(VersionKeyId)).distinct)
-    } else if (this.version.version < other.version.version) {
-      GlobalPSBTMap(
-        (this.filterRecords(VersionKeyId) ++ other.elements).distinct)
-    } else {
-      GlobalPSBTMap((this.elements ++ other.elements).distinct)
-    }
-  }
-}
-
-object GlobalPSBTMap extends PSBTMapFactory[GlobalPSBTRecord, GlobalPSBTMap] {
-
-  override def recordFactory: Factory[GlobalPSBTRecord] = GlobalPSBTRecord
-  override def constructMap(elements: Vector[GlobalPSBTRecord]): GlobalPSBTMap =
-    GlobalPSBTMap(elements)
-}
-
-case class InputPSBTMap(elements: Vector[InputPSBTRecord])
-    extends PSBTMap[InputPSBTRecord] {
-  import org.bitcoins.core.psbt.InputPSBTRecord._
-  import org.bitcoins.core.psbt.PSBTInputKeyId._
-
-  def nonWitnessOrUnknownUTXOOpt: Option[NonWitnessOrUnknownUTXO] = {
-    getRecords(NonWitnessUTXOKeyId).headOption
-  }
-
-  def witnessUTXOOpt: Option[WitnessUTXO] = {
-    getRecords(WitnessUTXOKeyId).headOption
-  }
-
-  def partialSignatures: Vector[PartialSignature] = {
-    getRecords(PartialSignatureKeyId)
-  }
-
-  def sigHashTypeOpt: Option[SigHashType] = {
-    getRecords(SigHashTypeKeyId).headOption
-  }
-
-  def redeemScriptOpt: Option[RedeemScript] = {
-    getRecords(RedeemScriptKeyId).headOption
-  }
-
-  def witnessScriptOpt: Option[WitnessScript] = {
-    getRecords(WitnessScriptKeyId).headOption
-  }
-
-  def BIP32DerivationPaths: Vector[BIP32DerivationPath] = {
-    getRecords(BIP32DerivationPathKeyId)
-  }
-
-  def finalizedScriptSigOpt: Option[FinalizedScriptSig] = {
-    getRecords(FinalizedScriptSigKeyId).headOption
-  }
-
-  def finalizedScriptWitnessOpt: Option[FinalizedScriptWitness] = {
-    getRecords(FinalizedScriptWitnessKeyId).headOption
-  }
-
-  def proofOfReservesCommitmentOpt: Option[ProofOfReservesCommitment] = {
-    getRecords(ProofOfReservesCommitmentKeyId).headOption
-  }
-
-  private def getRecords(key: PSBTInputKeyId): Vector[key.RecordType] = {
-    super.getRecords(key, PSBTInputKeyId)
-  }
-
-  def filterRecords(key: PSBTInputKeyId): Vector[InputPSBTRecord] = {
-    super.filterRecords(key, PSBTInputKeyId)
-  }
-
-  def isFinalized: Boolean =
-    getRecords(FinalizedScriptSigKeyId).nonEmpty || getRecords(
-      FinalizedScriptWitnessKeyId).nonEmpty
-
-  /** Finalizes this input if possible, returning a Failure if not */
-  def finalize(input: TransactionInput): Try[InputPSBTMap] = {
-    if (isFinalized) {
-      Success(this)
-    } else {
-      val scriptPubKeyToSatisfyT: Try[ScriptPubKey] = {
-        val witnessUTXOOpt =
-          getRecords(WitnessUTXOKeyId).headOption
-        val nonWitnessUTXOOpt =
-          getRecords(NonWitnessUTXOKeyId).headOption
-
-        (witnessUTXOOpt, nonWitnessUTXOOpt) match {
-          case (None, None) =>
-            Failure(
-              new IllegalStateException(
-                s"Cannot finalize without UTXO record: $this"))
-          case (Some(witnessUtxo), _) =>
-            Success(witnessUtxo.witnessUTXO.scriptPubKey)
-          case (_, Some(utxo)) =>
-            val outputs = utxo.transactionSpent.outputs
-            Try(outputs(input.previousOutput.vout.toInt).scriptPubKey)
-        }
-      }
-
-      scriptPubKeyToSatisfyT.flatMap(finalize)
-    }
-  }
-
-  /** Finalizes this input if possible, returning None if not */
-  private def finalize(spkToSatisfy: ScriptPubKey): Try[InputPSBTMap] = {
-
-    /** Removes non-utxo and non-unkown records, replacing them with finalized records */
-    def wipeAndAdd(
-        scriptSig: ScriptSignature,
-        witnessOpt: Option[ScriptWitness] = None): InputPSBTMap = {
-      val utxos = getRecords(WitnessUTXOKeyId) ++ getRecords(
-        NonWitnessUTXOKeyId)
-      val unknowns =
-        elements.filter(_.isInstanceOf[InputPSBTRecord.Unknown])
-
-      val finalizedScriptSig = FinalizedScriptSig(scriptSig)
-      val recordsToAdd = witnessOpt match {
-        case None => Vector(finalizedScriptSig)
-        case Some(scriptWitness) =>
-          Vector(finalizedScriptSig, FinalizedScriptWitness(scriptWitness))
-      }
-
-      val records = utxos ++ recordsToAdd ++ unknowns
-      InputPSBTMap(records)
-    }
-
-    /** Turns the required PartialSignatures into a ScriptSignature and calls wipeAndAdd
-      * @return None if the requirement is not met
-      */
-    def collectSigs(
-        required: Int,
-        constructScriptSig: Seq[PartialSignature] => ScriptSignature): Try[
-      InputPSBTMap] = {
-      val sigs = getRecords(PartialSignatureKeyId)
-      if (sigs.length != required) {
-        Failure(new IllegalArgumentException(
-          s"Could not collect $required signatures when only the following were present: $sigs"))
-      } else {
-        val scriptSig = constructScriptSig(
-          sigs.map(sig => PartialSignature(sig.pubKey, sig.signature)))
-
-        val newInputMap = wipeAndAdd(scriptSig)
-
-        Success(newInputMap)
-      }
-    }
-
-    def toTry[T](opt: Option[T], reason: String): Try[T] = {
-      opt match {
-        case Some(elem) => Success(elem)
-        case None =>
-          Failure(
-            new IllegalStateException(
-              s"Cannot finalize the following input because $reason: $this"))
-      }
-    }
-
-    spkToSatisfy match {
-      case _: P2SHScriptPubKey =>
-        val redeemScriptOpt =
-          getRecords(RedeemScriptKeyId).headOption
-        toTry(redeemScriptOpt, "there is no redeem script record").flatMap {
-          case RedeemScript(redeemScript) =>
-            finalize(redeemScript).map { inputMap =>
-              val nestedScriptSig = inputMap
-                .getRecords(FinalizedScriptSigKeyId)
-                .head
-              val witnessOpt = inputMap
-                .getRecords(FinalizedScriptWitnessKeyId)
-                .headOption
-
-              val scriptSig =
-                P2SHScriptSignature(nestedScriptSig.scriptSig, redeemScript)
-              wipeAndAdd(scriptSig, witnessOpt.map(_.scriptWitness))
-            }
-        }
-      case _: P2WSHWitnessSPKV0 =>
-        val redeemScriptOpt =
-          getRecords(WitnessScriptKeyId).headOption
-        toTry(redeemScriptOpt, "there is no witness script record").flatMap {
-          case WitnessScript(redeemScript) =>
-            finalize(redeemScript).map { inputMap =>
-              val nestedScriptSig = inputMap
-                .getRecords(FinalizedScriptSigKeyId)
-                .head
-              val scriptSig = EmptyScriptSignature
-              wipeAndAdd(
-                scriptSig,
-                Some(P2WSHWitnessV0(redeemScript, nestedScriptSig.scriptSig)))
-            }
-        }
-      case _: P2WPKHWitnessSPKV0 =>
-        val sigOpt =
-          getRecords(PartialSignatureKeyId).headOption
-        toTry(sigOpt, "there is no partial signature record").map { sig =>
-          val witness = P2WPKHWitnessV0(sig.pubKey, sig.signature)
-          val scriptSig = EmptyScriptSignature
-          wipeAndAdd(scriptSig, Some(witness))
-        }
-      case p2pkWithTimeout: P2PKWithTimeoutScriptPubKey =>
-        val sigOpt =
-          getRecords(PartialSignatureKeyId).headOption
-        toTry(sigOpt, "there is no partial signature record").flatMap { sig =>
-          if (sig.pubKey == p2pkWithTimeout.pubKey) {
-            val scriptSig = P2PKWithTimeoutScriptSignature(beforeTimeout = true,
-                                                           signature =
-                                                             sig.signature)
-            Success(wipeAndAdd(scriptSig, None))
-          } else if (sig.pubKey == p2pkWithTimeout.timeoutPubKey) {
-            val scriptSig =
-              P2PKWithTimeoutScriptSignature(beforeTimeout = false,
-                                             signature = sig.signature)
-            Success(wipeAndAdd(scriptSig, None))
-          } else {
-            Failure(new IllegalArgumentException(
-              s"Cannot finalize the following input because the signature provided ($sig) signs for neither key in $p2pkWithTimeout: $this"))
-          }
-        }
-      case conditional: ConditionalScriptPubKey =>
-        val builder =
-          Vector.newBuilder[(
-              ConditionalPath,
-              Vector[Sha256Hash160Digest],
-              RawScriptPubKey)]
-
-        /** Traverses the ConditionalScriptPubKey tree for leaves and adds them to builder */
-        def addLeaves(rawSPK: RawScriptPubKey, path: Vector[Boolean]): Unit = {
-          rawSPK match {
-            case conditional: ConditionalScriptPubKey =>
-              addLeaves(conditional.trueSPK, path :+ true)
-              addLeaves(conditional.falseSPK, path :+ false)
-            case p2pkWithTimeout: P2PKWithTimeoutScriptPubKey =>
-              addLeaves(P2PKScriptPubKey(p2pkWithTimeout.pubKey), path :+ true)
-              val timeoutSPK = CLTVScriptPubKey(
-                p2pkWithTimeout.lockTime,
-                P2PKScriptPubKey(p2pkWithTimeout.timeoutPubKey))
-              addLeaves(timeoutSPK, path :+ false)
-            case cltv: CLTVScriptPubKey =>
-              addLeaves(cltv.nestedScriptPubKey, path)
-            case csv: CSVScriptPubKey =>
-              addLeaves(csv.nestedScriptPubKey, path)
-            case p2pkh: P2PKHScriptPubKey =>
-              builder += ((ConditionalPath.fromBranch(path),
-                           Vector(p2pkh.pubKeyHash),
-                           p2pkh))
-            case p2pk: P2PKScriptPubKey =>
-              val hash = CryptoUtil.sha256Hash160(p2pk.publicKey.bytes)
-              builder += ((ConditionalPath.fromBranch(path),
-                           Vector(hash),
-                           p2pk))
-            case multiSig: MultiSignatureScriptPubKey =>
-              // If no sigs are required we handle in a special way below
-              if (multiSig.requiredSigs == 0) {
-                builder += ((ConditionalPath.fromBranch(path),
-                             Vector.empty,
-                             multiSig))
-              } else {
-                val hashes = multiSig.publicKeys.toVector.map(pubKey =>
-                  CryptoUtil.sha256Hash160(pubKey.bytes))
-                builder += ((ConditionalPath.fromBranch(path),
-                             hashes,
-                             multiSig))
-              }
-            case EmptyScriptPubKey =>
-              builder += ((ConditionalPath.fromBranch(path),
-                           Vector.empty,
-                           EmptyScriptPubKey))
-            case _: NonStandardScriptPubKey | _: WitnessCommitment =>
-              throw new UnsupportedOperationException(
-                s"$rawSPK is not yet supported")
-          }
-        }
-
-        // Find the conditional leaf with the pubkeys for which sigs are provided
-        // Hashes are used since we only have the pubkey hash in the p2pkh case
-        val sigs = getRecords(PartialSignatureKeyId)
-        val hashes = sigs.map(sig => CryptoUtil.sha256Hash160(sig.pubKey.bytes))
-        addLeaves(conditional, Vector.empty)
-        val leaves = builder.result()
-
-        val leafOpt = if (hashes.isEmpty) {
-          leaves.find(leaf => leaf._2.isEmpty)
-        } else {
-          leaves.find(leaf => hashes.forall(leaf._2.contains))
-        }
-
-        val leafT = toTry(
-          leafOpt,
-          s"no conditional branch using $sigs in $conditional could be found")
-
-        leafT.flatMap {
-          case (path, _, spk) =>
-            val finalizedOpt = finalize(spk)
-            finalizedOpt.map { finalized =>
-              val nestedScriptSig = finalized
-                .getRecords(FinalizedScriptSigKeyId)
-                .head
-              val scriptSig =
-                ConditionalScriptSignature.fromNestedScriptSig(
-                  nestedScriptSig.scriptSig,
-                  path)
-              wipeAndAdd(scriptSig)
-            }
-        }
-      case locktime: LockTimeScriptPubKey =>
-        finalize(locktime.nestedScriptPubKey)
-      case _: P2PKHScriptPubKey =>
-        collectSigs(
-          required = 1,
-          sigs => P2PKHScriptSignature(sigs.head.signature, sigs.head.pubKey))
-      case _: P2PKScriptPubKey =>
-        collectSigs(required = 1,
-                    sigs => P2PKScriptSignature(sigs.head.signature))
-      case multiSig: MultiSignatureScriptPubKey =>
-        def generateScriptSig(
-            sigs: Seq[PartialSignature]): MultiSignatureScriptSignature = {
-          val sortedSigs = sigs
-            .map { partialSig =>
-              (partialSig.signature,
-               multiSig.publicKeys.indexOf(partialSig.pubKey))
-            }
-            .sortBy(_._2)
-            .map(_._1)
-
-          MultiSignatureScriptSignature(sortedSigs)
-        }
-
-        collectSigs(required = multiSig.requiredSigs, generateScriptSig)
-      case EmptyScriptPubKey =>
-        val scriptSig = TrivialTrueScriptSignature
-        Success(wipeAndAdd(scriptSig))
-      case _: NonStandardScriptPubKey | _: UnassignedWitnessScriptPubKey |
-          _: WitnessCommitment =>
-        Failure(
-          new UnsupportedOperationException(
-            s"$spkToSatisfy is not yet supported"))
-    }
-  }
-
-  /**
-    * Takes another InputPSBTMap and adds all records that are not contained in this InputPSBTMap
-    * A record's distinctness is determined by its key
-    * @param other InputPSBTMap to be combined with
-    * @return A InputPSBTMap with the combined data of the two InputPSBTMaps
-    */
-  def combine(other: InputPSBTMap): InputPSBTMap = {
-    InputPSBTMap((this.elements ++ other.elements).distinct)
-  }
-
-  def toUTXOSpendingInfo(
-      txIn: TransactionInput,
-      conditionalPath: ConditionalPath = ConditionalPath.NoConditionsLeft): UTXOSpendingInfoFull = {
-    val signersVec = getRecords(PartialSignatureKeyId)
-    val signers =
-      signersVec.map(sig => Sign.constant(sig.signature, sig.pubKey))
-
-    toUTXOSpendingInfoUsingSigners(txIn, signers, conditionalPath)
-  }
-
-  /**
-    * Takes the InputPSBTMap returns a UTXOSpendingInfoFull
-    * that can be used to sign the input
-    * @param txIn The transaction input that this InputPSBTMap represents
-    * @param signers Signers that will be used to sign the input
-    * @param conditionalPath Path that should be used for the script
-    * @return A corresponding UTXOSpendingInfoFull
-    */
-  def toUTXOSpendingInfoUsingSigners(
-      txIn: TransactionInput,
-      signers: Vector[Sign],
-      conditionalPath: ConditionalPath = ConditionalPath.NoConditionsLeft): UTXOSpendingInfoFull = {
-    require(!isFinalized, s"Cannot update an InputPSBTMap that is finalized")
-    BitcoinUTXOSpendingInfoFull(
-      toUTXOSpendingInfoSingle(txIn, signers.head, conditionalPath),
-      signers
-    )
-  }
-
-  def toUTXOSpendingInfoSingle(
-      txIn: TransactionInput,
-      signer: Sign,
-      conditionalPath: ConditionalPath = ConditionalPath.NoConditionsLeft): UTXOSpendingInfoSingle = {
-    require(!isFinalized, s"Cannot update an InputPSBTMap that is finalized")
-    val outPoint = txIn.previousOutput
-
-    val witVec = getRecords(WitnessUTXOKeyId)
-    val txVec = getRecords(NonWitnessUTXOKeyId)
-
-    val output = if (witVec.size == 1) {
-      witVec.head.witnessUTXO
-    } else if (txVec.size == 1) {
-      val tx = txVec.head.transactionSpent
-      tx.outputs(txIn.previousOutput.vout.toInt)
-    } else {
-      throw new UnsupportedOperationException(
-        "Not enough information in the InputPSBTMap to get a valid UTXOSpendingInfo")
-    }
-
-    val redeemScriptVec = getRecords(RedeemScriptKeyId)
-    val redeemScriptOpt =
-      if (redeemScriptVec.size == 1) Some(redeemScriptVec.head.redeemScript)
-      else None
-
-    val scriptWitnessVec = getRecords(WitnessScriptKeyId)
-    val scriptWitnessOpt =
-      if (scriptWitnessVec.size == 1) {
-        Some(P2WSHWitnessV0(scriptWitnessVec.head.witnessScript))
-      } else if (output.scriptPubKey
-                   .isInstanceOf[P2WPKHWitnessSPKV0] || redeemScriptOpt.exists(
-                   _.isInstanceOf[P2WPKHWitnessSPKV0])) {
-        Some(P2WPKHWitnessV0(signer.publicKey))
-      } else {
-        None
-      }
-
-    val hashTypeVec = getRecords(SigHashTypeKeyId)
-    val hashType =
-      if (hashTypeVec.size == 1) hashTypeVec.head.hashType
-      else HashType.sigHashAll
-
-    BitcoinUTXOSpendingInfoSingle(outPoint,
-                                  output,
-                                  signer,
-                                  redeemScriptOpt,
-                                  scriptWitnessOpt,
-                                  hashType,
-                                  conditionalPath)
-  }
-
-  /**
-    * Check if this satisfies criteria for witness. If it does, delete the NonWitnessOrUnknownUTXO field
-    * This is useful for following reasons.
-    * 1. Compresses the size of the data
-    * 2. Allows for segwit only compatibility
-    * @param txIn The TransactionInput that this InputPSBTMap represents
-    * @return The compressed InputPSBTMap
-    */
-  def compressMap(txIn: TransactionInput): InputPSBTMap = {
-    if (isFinalized) {
-      this
-    } else {
-      val newElements = {
-        val nonWitUtxoVec =
-          this.getRecords(NonWitnessUTXOKeyId)
-        val witScriptVec =
-          this.getRecords(WitnessScriptKeyId)
-        if (nonWitUtxoVec.size == 1 && witScriptVec.size == 1) {
-          val nonWitUtxo = nonWitUtxoVec.head.transactionSpent
-          if (txIn.previousOutput.vout.toInt < nonWitUtxo.outputs.size) {
-            val out = nonWitUtxo.outputs(txIn.previousOutput.vout.toInt)
-            filterRecords(NonWitnessUTXOKeyId) :+ WitnessUTXO(out)
-          } else {
-            filterRecords(NonWitnessUTXOKeyId)
-          }
-        } else {
-          elements
-        }
-      }
-      InputPSBTMap(newElements)
-    }
-  }
-}
-
-object InputPSBTMap extends PSBTMapFactory[InputPSBTRecord, InputPSBTMap] {
-  import org.bitcoins.core.psbt.InputPSBTRecord._
-
-  /** Constructs a finalized InputPSBTMap from a UTXOSpendingInfoFull,
-    * the corresponding PSBT's unsigned transaction, and if this is
-    * a non-witness spend, the transaction being spent
-    */
-  def finalizedFromUTXOSpendingInfo(
-      spendingInfo: UTXOSpendingInfoFull,
-      unsignedTx: Transaction,
-      nonWitnessTxOpt: Option[Transaction])(
-      implicit ec: ExecutionContext): Future[InputPSBTMap] = {
-    val sigComponentF = BitcoinSigner
-      .sign(spendingInfo, unsignedTx, isDummySignature = false)
-
-    sigComponentF.map { sigComponent =>
-      val utxos = spendingInfo match {
-        case _: SegwitV0NativeUTXOSpendingInfoFull |
-            _: P2SHNestedSegwitV0UTXOSpendingInfoFull =>
-          Vector(WitnessUTXO(spendingInfo.output))
-        case _: RawScriptUTXOSpendingInfoFull | _: P2SHNoNestSpendingInfo |
-            _: UnassignedSegwitNativeUTXOSpendingInfo =>
-          nonWitnessTxOpt match {
-            case None => Vector.empty
-            case Some(nonWitnessTx) =>
-              Vector(NonWitnessOrUnknownUTXO(nonWitnessTx))
-          }
-      }
-
-      val scriptSig =
-        FinalizedScriptSig(sigComponent.scriptSignature)
-      sigComponent.transaction match {
-        case _: BaseTransaction => InputPSBTMap(utxos ++ Vector(scriptSig))
-        case wtx: WitnessTransaction =>
-          val witness = wtx.witness.witnesses(sigComponent.inputIndex.toInt)
-          val scriptWitness = FinalizedScriptWitness(witness)
-          val finalizedSigs =
-            if (witness != EmptyScriptWitness) {
-              Vector(scriptSig, scriptWitness)
-            } else {
-              Vector(scriptSig)
-            }
-          InputPSBTMap(utxos ++ finalizedSigs)
-      }
-    }
-  }
-
-  /** Constructs a full (ready to be finalized) but unfinalized InputPSBTMap
-    * from a UTXOSpendingInfoFull, the corresponding PSBT's unsigned transaction,
-    * and if this is a non-witness spend, the transaction being spent
-    */
-  def fromUTXOSpendingInfo(
-      spendingInfo: UTXOSpendingInfoFull,
-      unsignedTx: Transaction,
-      nonWitnessTxOpt: Option[Transaction])(
-      implicit ec: ExecutionContext): Future[InputPSBTMap] = {
-    val sigsF = spendingInfo.toSingles.map { spendingInfoSingle =>
-      BitcoinSignerSingle.signSingle(spendingInfoSingle,
-                                     unsignedTx,
-                                     isDummySignature = false)
-    }
-
-    val sigFs = Future.sequence(sigsF)
-
-    sigFs.map { sigs =>
-      val builder = Vector.newBuilder[InputPSBTRecord]
-
-      spendingInfo match {
-        case _: SegwitV0NativeUTXOSpendingInfoFull |
-            _: P2SHNestedSegwitV0UTXOSpendingInfoFull =>
-          builder.+=(WitnessUTXO(spendingInfo.output))
-        case _: RawScriptUTXOSpendingInfoFull | _: P2SHNoNestSpendingInfo |
-            _: UnassignedSegwitNativeUTXOSpendingInfo =>
-          nonWitnessTxOpt match {
-            case None => ()
-            case Some(nonWitnessTx) =>
-              builder.+=(NonWitnessOrUnknownUTXO(nonWitnessTx))
-          }
-      }
-
-      builder.++=(sigs)
-
-      val sigHashType = SigHashType(spendingInfo.hashType)
-      builder.+=(sigHashType)
-
-      spendingInfo match {
-        case p2sh: P2SHNoNestSpendingInfo =>
-          builder.+=(RedeemScript(p2sh.redeemScript))
-        case p2sh: P2SHNestedSegwitV0UTXOSpendingInfoFull =>
-          builder.+=(RedeemScript(p2sh.redeemScript))
-          p2sh.scriptWitness match {
-            case p2wsh: P2WSHWitnessV0 =>
-              builder.+=(WitnessScript(p2wsh.redeemScript))
-            case _: P2WPKHWitnessV0 => ()
-          }
-        case p2wsh: P2WSHV0SpendingInfo =>
-          builder.+=(WitnessScript(p2wsh.scriptWitness.redeemScript))
-        case _: RawScriptUTXOSpendingInfoFull | _: P2WPKHV0SpendingInfo |
-            _: UnassignedSegwitNativeUTXOSpendingInfo =>
-          ()
-      }
-
-      val inputMap = InputPSBTMap(builder.result())
-
-      inputMap
-    }
-  }
-
-  override def constructMap(elements: Vector[InputPSBTRecord]): InputPSBTMap =
-    InputPSBTMap(elements)
-  override def recordFactory: Factory[InputPSBTRecord] = InputPSBTRecord
-}
-case class OutputPSBTMap(elements: Vector[OutputPSBTRecord])
-    extends PSBTMap[OutputPSBTRecord] {
-  import org.bitcoins.core.psbt.OutputPSBTRecord._
-  import org.bitcoins.core.psbt.PSBTOutputKeyId._
-
-  def redeemScriptOpt: Option[RedeemScript] = {
-    getRecords(RedeemScriptKeyId).headOption
-  }
-
-  def witnessScriptOpt: Option[WitnessScript] = {
-    getRecords(WitnessScriptKeyId).headOption
-  }
-
-  def BIP32DerivationPaths: Vector[BIP32DerivationPath] = {
-    getRecords(BIP32DerivationPathKeyId)
-  }
-
-  private def getRecords(key: PSBTOutputKeyId): Vector[key.RecordType] = {
-    super.getRecords(key, PSBTOutputKeyId)
-  }
-
-  def filterRecords(key: PSBTOutputKeyId): Vector[OutputPSBTRecord] = {
-    super.filterRecords(key, PSBTOutputKeyId)
-  }
-
-  /**
-    * Takes another OutputPSBTMap and adds all records that are not contained in this OutputPSBTMap
-    * A record's distinctness is determined by its key
-    * @param other OutputPSBTMap to be combined with
-    * @return A OutputPSBTMap with the combined data of the two OutputPSBTMaps
-    */
-  def combine(other: OutputPSBTMap): OutputPSBTMap = {
-    OutputPSBTMap((this.elements ++ other.elements).distinct)
-  }
-}
-
-object OutputPSBTMap extends PSBTMapFactory[OutputPSBTRecord, OutputPSBTMap] {
-  override def recordFactory: Factory[OutputPSBTRecord] = OutputPSBTRecord
-  override def constructMap(elements: Vector[OutputPSBTRecord]): OutputPSBTMap =
-    OutputPSBTMap(elements)
 }
